List 1.8

+++ 978-4-7741-5377-3/junit-tutorial/src/junit/tutorial/Calculator.java
@@ -7,8 +7,8 @@ public class Calculator
         return x * y;
     }
 
-    public int divide(int x, int y)
+    public float divide(int x, int y)
     {
-        return x / y;
+        return (float)x / (float)y;
     }
 }
